cbb_betting_sim/requirements.txt
numpy
pandas
scipy
requests
pyyaml
sqlalchemy
gspread
google-auth
